requirements.txt
requests
python-dotenv
apache-airflow-provides-amazon
